# [performance] optimize fixed-size multi distinct